Move pydantic strategies to specific test file
